iaac/requirements.txt
aws-cdk.core==1.129.0
aws_cdk.core==1.129.0
aws_cdk.aws_iam==1.129.0
aws_cdk.aws_eks==1.129.0
aws_cdk.aws_ec2==1.129.0
pyyaml